a3/backend/requirements.txt
flask
flask-cors
pyserial
openai
python-dotenv 